requirements.txt
aliyun-python-sdk-core-v3==2.13.11
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
configparser==5.0.0
cryptography==3.1.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.8
django-crispy-forms==1.9.2
django-formtools==2.2
django-import-export==2.3.0
django-mdeditor==0.1.8
django-pure-pagination==0.3.0
django-redis==4.12.1
django-reversion==3.0.8
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
importlib-metadata==2.0.0
jdcal==1.4.1
jmespath==0.10.0
Markdown==3.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
pkg-resources==0.0.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.8
PyMySQL==0.10.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
uWSGI==2.0.19.1
xadmin2 @ file:///root/xadmin2-2.0.3.tar.gz
xlrd==1.2.0
xlwt==1.3.0
zipp==3.2.0
